pyproject.toml
[project]
name = "import-hawkeye"
version = "0.1.0"
description = "Unified Hawkeye Import Cloud Function - consolidates VHF, GPS, Airdefense, Skytrace, and Shadowfleet imports"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pendulum>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "functions-framework>=3.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 140

[tool.ruff.lint]
extend-select = ["PL"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
